Grid Script Stub
Working on a resource instance for the grid system (thank you gdquest) so that all nodes that need it can have it as a reference. The grid currently has the dimension defined and cell size. The grid will be centered on the origin of the tilemap and will only map over the game board (the tiles with circles on them).